Tweak Postgres migration script.

+++ modules/module-postgres-storage/src/migrations/scripts/1771232439485-storage-version.ts
@@ -9,15 +9,7 @@ export const up: migrations.PowerSyncMigrationFunction = async (context) => {
   await client.transaction(async (db) => {
     await db.sql`
       ALTER TABLE sync_rules
-      ADD COLUMN storage_version integer
-    `.execute();
-
-    await db.sql`
-      UPDATE sync_rules
-      SET
-        storage_version = ${{ type: 'int4', value: storage.LEGACY_STORAGE_VERSION }}
-      WHERE
-        storage_version IS NULL
+      ADD COLUMN storage_version integer NOT NULL DEFAULT 1
     `.execute();
   });
 };
@@ -28,6 +20,22 @@ export const down: migrations.PowerSyncMigrationFunction = async (context) => {
   } = context;
   await using client = openMigrationDB(configuration.storage);
   await client.transaction(async (db) => {
+    const newRules = await db.sql`
+      SELECT
+        id,
+        storage_version
+      FROM
+        sync_rules
+      WHERE
+        storage_version > ${{ type: 'int4', value: storage.LEGACY_STORAGE_VERSION }}
+    `.rows<{ id: number | bigint; storage_version: number | bigint }>();
+
+    if (newRules.length > 0) {
+      throw new Error(
+        `Cannot revert migration due to newer storage versions in use: ${newRules.map((r) => `${r.id}: v${r.storage_version}`).join(', ')}`
+      );
+    }
+
     await db.sql`
       ALTER TABLE sync_rules
       DROP COLUMN storage_version
